application/external_apps/mcp_server_auth/requirements.txt
fastmcp>=2.10.0
msal>=1.33.0
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
cryptography>=41.0.0
PyJWT>=2.8.0
nest-asyncio>=1.5.0
